Fix bug with events.

+++ bot_manager/services/helpers/condition_parser.py
@@ -148,8 +148,8 @@ class ConditionParser:
         cost = 0.0
         leads = 0
 
-        event_2 = 0
-        event_5 = 0
+        event_2 = 0.0
+        event_5 = 0.0
 
         for offer_stat in statistics:
             clicks += int(offer_stat['clicks'])
@@ -157,8 +157,8 @@ class ConditionParser:
             revenue += float(offer_stat['revenue'])
             cost += float(offer_stat['cost'])
             leads += int(offer_stat['leads'])
-            event_2 += int(offer_stat['event_2'])
-            event_5 = int(offer_stat['event_5'])
+            event_2 += float(offer_stat['event_2'])
+            event_5 = float(offer_stat['event_5'])
 
         if var == 'clicks':
             var = clicks
